LayoutTests/http/tests/xmlhttprequest/access-control-basic-non-simple-deny-cached-expected.txt

PASS: Exception thrown. Cross-domain access is not allowed in 'send'. [NETWORK_ERR: XMLHttpRequest Exception 101].
